RUE-858: split aggregate semantic analysis
